requirements.txt
streamlit>=1.37
azure-storage-blob>=12.19.0
azure-ai-translation-document>=1.0.0b1
python-dotenv>=1.0.0
requests>=2.32
python-pptx>=0.6.21
azure-search-documents>=11.4.0
azure-identity>=1.15.0
openai>=1.0.0
azure-ai-documentintelligence>=1.0.0b1
pandas
openpyxl
extra-streamlit-components
